+++ rest/src/index.ts
@@ -4,5 +4,4 @@ export * from './middlewares';
 export * from './misc/defaults';
 export type {middlewareTypes} from './misc/middlewareTypes';
 export * from './misc/restHttpError.interface';
-export * from './misc/tokens';
 export * from './misc/withRestClientContext';